+++ asdf/extension.py
@@ -9,10 +9,11 @@ from pkg_resources import iter_entry_points
 import six
 import importlib
 
-from . import asdftypes
+from . import types
 from . import resolver
-from .version import version as asdf_version
 from .util import get_class_name
+from .type_index import AsdfTypeIndex
+from .version import version as asdf_version
 from .exceptions import AsdfDeprecationWarning
 
 
@@ -120,11 +121,11 @@ class AsdfExtensionList(object):
         tag_mapping = []
         url_mapping = []
         validators = {}
-        self._type_index = asdftypes.AsdfTypeIndex()
+        self._type_index = AsdfTypeIndex()
         for extension in extensions:
             if not isinstance(extension, AsdfExtension):
                 raise TypeError(
-                    "Extension must implement asdftypes.AsdfExtension "
+                    "Extension must implement asdf.types.AsdfExtension "
                     "interface")
             tag_mapping.extend(extension.tag_mapping)
             url_mapping.extend(extension.url_mapping)
@@ -172,7 +173,7 @@ class BuiltinExtension(object):
     """
     @property
     def types(self):
-        return asdftypes._all_asdftypes
+        return types._all_asdftypes
 
     @property
     def tag_mapping(self):